+++ src/scripts/background.js
@@ -1,20 +1,19 @@
-import browser from 'webextension-polyfill'
-import { updateBlockingRules } from './comp/block'
-import { initParams } from './data/params'
+import browser from "webextension-polyfill";
+import { updateBlockingRules } from "./comp/block";
+import { initParams } from "./data/params";
 
 browser.runtime.onInstalled.addListener(async ({ reason, temporary }) => {
   // if (temporary) return; // skip during development
   switch (reason) {
-    case 'install':
-      await browser.runtime.openOptionsPage()
-      break
+    case "install":
+      await browser.runtime.openOptionsPage();
+      break;
   }
-})
+});
 
-browser.storage.onChanged.addListener((changes, areaName)=>{
-  if(areaName === 'local'){
-    updateBlockingRules()
+browser.storage.onChanged.addListener((changes, areaName) => {
+  if (areaName === "local") {
+    updateBlockingRules();
   }
-})
-initParams().then(() => {}, console.error)
-
+});
+initParams().then(() => {}, console.error);